+++ src/services/UpvoteService.php
@@ -13,6 +13,8 @@ namespace doublesecretagency\upvote\services;
 
 use Craft;
 use craft\base\Component;
+use craft\elements\User;
+use craft\helpers\Json;
 
 use doublesecretagency\upvote\Upvote;
 
@@ -29,6 +31,22 @@ class UpvoteService extends Component
     public $userCookieLifespan = 315569260; // Lasts 10 years
     public $anonymousHistory = [];
     public $loggedInHistory = [];
+    public $history;
+
+    //
+    public function init()
+    {
+        // If login is required
+        if (Upvote::$plugin->getSettings()->requireLogin) {
+            // Rely on user history from DB
+            $this->history =& $this->loggedInHistory;
+        } else {
+            // Rely on anonymous user history
+            $this->history =& $this->anonymousHistory;
+        }
+
+        parent::init();
+    }
 
     // Generate combined item key
     public function setItemKey($elementId, $key, $separator = ':')
@@ -79,7 +97,7 @@ class UpvoteService extends Component
         if ($cookies->has($this->userCookie)) {
             // Get anonymous history
             $cookieValue = $cookies->getValue($this->userCookie);
-            $this->anonymousHistory = json_decode($cookieValue, true);
+            $this->anonymousHistory = Json::decode($cookieValue);
         }
 
         // If no anonymous history
@@ -99,6 +117,89 @@ class UpvoteService extends Component
 
     // ========================================================================= //
 
+    /**
+     */
+    public function compileElementData($itemKey, $userVote = null, $isAntivote = false)
+    {
+        // Get current user
+        $currentUser = Craft::$app->user->getIdentity();
+
+        // Split ID into array
+        $parts = explode(':', $itemKey);
+
+        // Get the element ID
+        $elementId = (int) array_shift($parts);
+
+        // If no element ID, bail
+        if (!$elementId) {
+            return false;
+        }
+
+        // Reassemble the remaining parts (in case the key contains a colon)
+        $key = implode(':', $parts);
+
+        // If no key, set to null
+        if (!$key) {
+            $key = null;
+        }
+
+        // Get user's vote history for this item
+        $itemHistory = ($this->history[$itemKey] ?? null);
+
+        // Set vote configuration
+        $vote = [
+            'id' => $elementId,
+            'key' => $key,
+            'itemKey' => $itemKey,
+            'userId' => ($currentUser ? (int) $currentUser->id : null),
+            'userVote' => ($userVote ?? $itemHistory),
+            'isAntivote' => $isAntivote,
+        ];
+
+        // Get element totals from BEFORE the vote is calculated
+        $totals = [
+            'tally' => Upvote::$plugin->upvote_query->tally($elementId, $key),
+            'totalVotes' => Upvote::$plugin->upvote_query->totalVotes($elementId, $key),
+            'totalUpvotes' => Upvote::$plugin->upvote_query->totalUpvotes($elementId, $key),
+            'totalDownvotes' => Upvote::$plugin->upvote_query->totalDownvotes($elementId, $key),
+        ];
+
+        // If existing vote was removed
+        if ($isAntivote && $itemHistory) {
+            // Create antivote
+            $userVote = $itemHistory * -1;
+            // Set total type
+            $totalType = (1 === $userVote ? 'totalDownvotes' : 'totalUpvotes');
+        } else {
+            // Set total type
+            $totalType = (1 === $userVote ? 'totalUpvotes' : 'totalDownvotes');
+        }
+
+        // If a vote was cast or removed
+        if ($userVote) {
+
+            // Add to tally
+            $totals['tally'] += $userVote;
+
+            // If removing vote
+            if ($isAntivote) {
+                // One less vote
+                $totals['totalVotes']--;
+                $totals[$totalType]--;
+            } else {
+                // One more vote
+                $totals['totalVotes']++;
+                $totals[$totalType]++;
+            }
+
+        }
+
+        // Return element's vote data
+        return array_merge($vote, $totals);
+    }
+
+    // ========================================================================= //
+
     // $userId can be valid user ID or UserModel
     public function validateUserId(&$userId)
     {
@@ -113,7 +214,7 @@ class UpvoteService extends Component
             if (is_numeric($userId)) {
                 // Get valid UserModel
                 $user = Craft::$app->users->getUserById($userId);
-            } else if (is_object($userId) && is_a($userId, 'craft\\elements\\User')) {
+            } else if (is_object($userId) && is_a($userId, User::class)) {
                 // It's already a User model
                 $user = $userId;
             }